requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
matplotlib>=3.5.0
seaborn>=0.11.0
ta>=0.10.0
pyyaml>=6.0
jupyter>=1.0.0